+++ apisix/plugins/syslog/init.lua
@@ -18,7 +18,7 @@
 local core = require("apisix.core")
 local bp_manager_mod = require("apisix.utils.batch-processor-manager")
 local logger_socket = require("resty.logger.socket")
-local rfc5424 = require("apisix.syslog.rfc5424")
+local rfc5424 = require("apisix.utils.rfc5424")
 local ipairs = ipairs
 local table_insert = core.table.insert
 local table_concat = core.table.concat
